+++ renderer/pages/profile.js
@@ -27,6 +27,7 @@ import {
   InviteScoreAlert,
   KillIdentityDrawer,
   KillForm,
+  MyIdenaBotAlert,
 } from '../screens/profile/components'
 import {
   PrimaryButton,
@@ -214,6 +215,7 @@ export default function ProfilePage() {
     <>
       <InviteProvider>
         <Layout syncing={syncing} offline={offline}>
+          <MyIdenaBotAlert />
           <Page>
             <Stack spacing={8}>
               <Stack spacing={6}>
