+++ renderer/shared/components/layout.js
@@ -24,7 +24,6 @@ import {
   AlertDialogHeader,
   AlertDialogBody,
   AlertDialogFooter,
-  Checkbox,
 } from '@chakra-ui/core'
 import {useMachine} from '@xstate/react'
 import semver from 'semver'
@@ -80,7 +79,6 @@ import {
 } from '../../screens/dna/hooks'
 import {viewVotingHref} from '../../screens/oracles/utils'
 import {useFork} from '../../screens/hardfork/hooks'
-import {IdentityStatus} from '../types'
 
 global.getZoomLevel = global.getZoomLevel || {}
 
@@ -350,33 +348,8 @@ function NormalApp({children}) {
     }
   )
 
-  const [didConnectIdenaBot, setDidConnectIdenaBot] = React.useState(() => {
-    if (typeof window !== 'undefined') {
-      try {
-        return JSON.parse(localStorage.getItem('connectIdenaBot'))
-      } catch {
-        return null
-      }
-    }
-    return null
-  })
-
-  React.useEffect(() => {
-    if (typeof didConnectIdenaBot === 'boolean') {
-      localStorage.setItem('connectIdenaBot', didConnectIdenaBot)
-    }
-  }, [didConnectIdenaBot])
-
   return (
     <Flex as="section" direction="column" flex={1} h="100vh" overflowY="auto">
-      {!didConnectIdenaBot && (
-        <MyIdenaBotAlert
-          onConnect={() => {
-            setDidConnectIdenaBot(true)
-          }}
-        />
-      )}
-
       {children}
 
       {epoch && <ValidationToast epoch={epoch} identity={identity} />}
@@ -1072,152 +1045,5 @@ function ConfirmQuitDialog({onClose, onError, ...props}) {
         </AlertDialogFooter>
       </AlertDialogContent>
     </AlertDialog>
-  )
-}
-
-function MyIdenaBotAlert({onConnect}) {
-  const {t} = useTranslation()
-
-  const {state} = useIdentityState()
-
-  const myIdenaBotDisclosure = useDisclosure()
-
-  const [doNotShowAgain, setDoNotShowAgain] = React.useState()
-
-  // eslint-disable-next-line no-shadow
-  const eitherState = (...states) => states.some(s => s === state)
-
-  return (
-    <>
-      <Alert
-        variant="solid"
-        justifyContent="center"
-        flexShrink={0}
-        boxShadow="0 3px 12px 0 rgb(255 163 102 /0.1), 0 2px 3px 0 rgb(255 163 102 /0.2)"
-        color="white"
-        fontWeight={500}
-        fontSize="md"
-        rounded="md"
-        p={3}
-        m={2}
-        mb={0}
-        w="auto"
-        zIndex="banner"
-        cursor="pointer"
-        onClick={myIdenaBotDisclosure.onOpen}
-      >
-        {t(`Subscribe to @MyIdenaBot to get personalized notifications based on
-        your status`)}
-      </Alert>
-
-      <Dialog
-        title="Subscribe to @MyIdenaBot"
-        size="md"
-        {...myIdenaBotDisclosure}
-      >
-        <DialogBody>
-          <Stack>
-            <Text>
-              MyIdenaBot reminds you about important actions based on your
-              identity status:
-            </Text>
-
-            {eitherState(IdentityStatus.Undefined) && (
-              <IdenaBotFeatureList
-                features={[
-                  'next validation reminder',
-                  'notification when you get an invite',
-                  'reminder to activate your invite',
-                  'your validation results when validation consensus is reached',
-                ]}
-              />
-            )}
-
-            {eitherState(IdentityStatus.Invite, IdentityStatus.Candidate) && (
-              <IdenaBotFeatureList
-                features={[
-                  'next validation reminder',
-                  'your validation results when validation consensus is reached',
-                ]}
-              />
-            )}
-
-            {eitherState(IdentityStatus.Newbie) && (
-              <IdenaBotFeatureList
-                features={[
-                  'next validation reminder',
-                  'reminder to create flips if you haven’t done it yet and the validation is coming',
-                  'your validation results when validation consensus is reached',
-                ]}
-              />
-            )}
-
-            {eitherState(IdentityStatus.Verified, IdentityStatus.Human) && (
-              <IdenaBotFeatureList
-                features={[
-                  'next validation reminder',
-                  'reminder to create flips',
-                  'your validation results when validation consensus is reached',
-                  'reminder to share your remaining invites',
-                  'reminder to submit extra flips to get more rewards',
-                  'status update of all your invitees to check if they are ready for the validation (activated invites, submitted flips)',
-                ]}
-              />
-            )}
-            {eitherState(IdentityStatus.Zombie, IdentityStatus.Suspended) && (
-              <IdenaBotFeatureList
-                features={[
-                  'next validation reminder',
-                  'your validation results when validation consensus is reached',
-                  'reminder to share your remaining invites',
-                  'reminder to submit extra flips to get more rewards',
-                  'status update of all your invitees to check if they are ready for the validation (activated invites, submitted flips)',
-                ]}
-              />
-            )}
-          </Stack>
-        </DialogBody>
-        <DialogFooter align="center">
-          <Checkbox
-            borderColor="gray.100"
-            isChecked={doNotShowAgain}
-            onChange={e => {
-              setDoNotShowAgain(e.target.checked)
-            }}
-          >
-            {t('Do not show again')}
-          </Checkbox>
-          <SecondaryButton
-            onClick={() => {
-              myIdenaBotDisclosure.onClose()
-              if (doNotShowAgain) onConnect()
-            }}
-          >
-            {t('Not now')}
-          </SecondaryButton>
-          <PrimaryButton
-            onClick={() => {
-              global.openExternal('https://t.me/MyIdenaBot')
-              onConnect()
-            }}
-          >
-            {t('Connect')}
-          </PrimaryButton>
-        </DialogFooter>
-      </Dialog>
-    </>
-  )
-}
-
-function IdenaBotFeatureList({features, listSeparator = ';'}) {
-  return (
-    <List spacing={1} styleType="'- '">
-      {features.map((feature, idx) => (
-        <ListItem textTransform="lowercase">
-          {feature}
-          {idx < features.length - 1 ? listSeparator : '.'}
-        </ListItem>
-      ))}
-    </List>
   )
 }
